+++ src/pages/SonarFeed/ActivityRenderer/ActivityRenderer.module.scss
@@ -1,4 +1,4 @@
-@import "~@santiment-network/ui/mixins.scss";
+@import "../../../../node_modules/@santiment-network/ui/mixins";
 
 .description {
   margin-bottom: 12px;
