+++ src/pages/SonarFeed/SonarFeedActivityPage.js
@@ -6,7 +6,7 @@ import PageLoader from '../../components/Loader/PageLoader'
 import SonarFeedRecommendations from './SonarFeedRecommendations'
 import { getSanSonarSW } from '../Account/SettingsSonarWebPushNotifications'
 import { makeVariables } from '../feed/GeneralFeed/utils'
-import ActivityRenderer from './ActivityRenderer'
+import ActivityRenderer from './ActivityRenderer/ActivityRenderer'
 import { TRIGGERS_COMMON_FRAGMENT } from '../../ducks/Signals/common/queries'
 import styles from './SonarFeedActivityPage.module.scss'
